Netowrking automation
Author: Tregubov Yury

changed some function names and moved several methods in Enviroment
class
