+++ mos_tests/neutron/python_tests/base.py
@@ -140,49 +140,3 @@ class TestBase(object):
             wait(lambda: remote.execute(cmd)['exit_code'] == 0,
                  interval=10, timeout=3 * 10,
                  timeout_msg=error_msg)
-
-    def destroy_nodes(self, devops_nodes):
-        for node in devops_nodes:
-            node.destroy()
-        logger.info('wait untill the nodes get offline state')
-        assert(wait(lambda: self.wait_nodes_get_offline_state(
-                                 node_names=[x.name for x in devops_nodes]),
-                    timeout=10 * 60))
-        for node in self.env.get_all_nodes():
-            logger.info('online state of node {0} now is {1}'.
-                         format(node.data['name'], node.data['online']))
-
-    def warm_shutdown_nodes(self, devops_nodes):
-        for node in devops_nodes:
-            node_ip = node.get_ip_address_by_network_name('admin')
-            logger.info('Shutdown node {0} with ip {1}'.
-                         format(node.name, node_ip))
-            with self.env.get_ssh_to_node(node_ip) as remote:
-                remote.check_call('/sbin/shutdown -Ph now')
-        self.destroy_nodes(devops_nodes)
-
-    def warm_start_nodes(self, devops_nodes):
-        for node in devops_nodes:
-            logger.info('Starting node {}'.format(node.name))
-            node.create()
-        assert(wait(lambda: self.wait_nodes_get_online_state(),
-                    timeout=10 * 60))
-        logger.info('wait untill the nodes get online state')
-        for node in self.env.get_all_nodes():
-            logger.info('online state of node {0} now is {1}'.
-                         format(node.data['name'], node.data['online']))
-
-    def warm_restart_nodes(self, devops_nodes):
-        logger.info('Reboot (warm restart) nodes %s',
-                    [n.name for n in devops_nodes])
-        self.warm_shutdown_nodes(devops_nodes)
-        self.warm_start_nodes(devops_nodes)
-
-    def wait_nodes_get_offline_state(self, node_names=[]):
-        nodes_states = [not x.data['online']
-                        for x in self.env.get_all_nodes()
-                        if x.data['name'].split('_')[0] in node_names]
-        return all(nodes_states)
-
-    def wait_nodes_get_online_state(self):
-        return all([node.data['online'] for node in self.env.get_all_nodes()])
