+++ requirements.txt
@@ -4,7 +4,7 @@ pytest-xdist
 waiting
 python-glanceclient==0.17.1
 python-keystoneclient>=2.0.0
-python-novaclient>=2.15.0
+python-novaclient>=2.27.0
 python-cinderclient>=1.0.5
 python-neutronclient>=2.0
 python-heatclient>=0.2.12
